Move test data class in dir ignored by checkstyle
