+++ pmd-java/src/test/java/javasymbols/testdata/BrokenGeneric.java
@@ -1,9 +1,8 @@
-
 /*
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package javasymbols.brokenclasses;
+package javasymbols.testdata;
 
 import net.sourceforge.pmd.lang.java.symbols.internal.asm.Classpath;
